Batch study adaptations (#5)
* create segementation_classification

Add additional file as alternative to classification.js to additionally plot the probability output.
Copy based on classification.js from 04.04.25

* Delete segmentation_classification.js

* Automatic logout, userview-reset & neural-networkout panel

Upon reaching the end of a user's batch the user is now automatically logged out and userview is updated back to 1.

With new script double_users we create 2N users. For the first N users the probability output of the neural network is not shown, for the remaining N users it is shown. Users X and X+N have the exact same batches as vieworders.

* Update automatic logout & neural network panel

* Final update